+++ src/Cache/Delivery/CacheClearer.php
@@ -1,23 +1,25 @@
 <?php
 
 /**
- * This file is part of the ContentfulBundle package.
+ * This file is part of the contentful/contentful-bundle package.
  *
- * @copyright 2016-2018 Contentful GmbH
+ * @copyright 2015-2018 Contentful GmbH
  * @license   MIT
  */
 
-namespace Contentful\ContentfulBundle\Cache;
+declare(strict_types=1);
 
-use Contentful\Delivery\Cache\CacheClearer;
+namespace Contentful\ContentfulBundle\Cache\Delivery;
+
+use Contentful\Delivery\Cache\CacheClearer as SdkCacheClearer;
 use Contentful\Delivery\Client;
 use Psr\Cache\CacheItemPoolInterface;
 use Symfony\Component\HttpKernel\CacheClearer\CacheClearerInterface;
 
-class DeliveryCacheClearer implements CacheClearerInterface
+class CacheClearer implements CacheClearerInterface
 {
     /**
-     * @var CacheClearer
+     * @var SdkCacheClearer
      */
     private $clearer;
 
@@ -29,7 +31,7 @@ class DeliveryCacheClearer implements CacheClearerInterface
      */
     public function __construct(Client $client, CacheItemPoolInterface $cacheItemPool)
     {
-        $this->clearer = new CacheClearer($client, $cacheItemPool);
+        $this->clearer = new SdkCacheClearer($client, $client->getResourcePool(), $cacheItemPool);
     }
 
     /**